cleaned construction mechanism to only allow builder pattern to construct a separable model
