+++ src/model/basefunction.rs
@@ -8,7 +8,6 @@ use std::collections::{HashMap};
 
 
 use super::OwnedVector;
-use crate::model::errors::ModelfunctionError;
 
 pub struct Basefunction<ScalarType, NData>
     where ScalarType: Scalar,
@@ -36,42 +35,3 @@ impl<ScalarType, NData> Basefunction<ScalarType, NData>
         }
     }
 }
-
-/// This is a helper type that is not useful on its own. It only gets generated as part of
-/// pushing a function to a separable model.
-#[must_use="This alone will not push the function to the set. To do that, append a call to 'push()'."]
-pub struct ParameterIndepententModelFunctionProxy<'a, ScalarType, NData, FuncType>
-    where ScalarType: Scalar,
-          NData: Dim ,
-          nalgebra::DefaultAllocator: nalgebra::allocator::Allocator<ScalarType, NData>,
-          FuncType: Fn(&OwnedVector<ScalarType, NData>) -> OwnedVector<ScalarType, NData> + 'static
-{
-    model: &'a mut SeparableModel<ScalarType, NData>,
-    function: FuncType,
-}
-
-impl<'a, ScalarType, NData, FuncType> ParameterIndepententModelFunctionProxy<'a, ScalarType, NData, FuncType>
-    where ScalarType: Scalar,
-          NData: Dim ,
-          nalgebra::DefaultAllocator: nalgebra::allocator::Allocator<ScalarType, NData>,
-          FuncType: Fn(&OwnedVector<ScalarType, NData>) -> OwnedVector<ScalarType, NData> +'static {
-
-    pub(in crate::model) fn new(model : &'a mut SeparableModel<ScalarType, NData>, function : FuncType) -> Self {
-        Self {
-            model,
-            function
-        }
-    }
-
-    /// push this function to the set of model functions. `push` is
-    /// used after the call to [new_independent_function] of [SeparableModel] to push the function
-    /// into the model. `push` always returns with an Ok-variant but the interface is
-    /// the same as for parameter dependent functions, which might return an error variant.
-    pub fn push(self)->Result<(),ModelfunctionError> {
-        let Self{ model, function } = self;
-        model.modelfunctions.push(Basefunction::parameter_independent(move |x , _model_params| (function)(x)));
-        Ok(())
-    }
-}
-
-
