+++ src/model/errors.rs
@@ -4,7 +4,7 @@ use snafu::Snafu;
 /// the model and the set of model
 #[derive(Debug, Clone, Snafu, PartialEq)]
 #[snafu(visibility = "pub")]
-pub enum ModelfunctionError {
+pub enum ModelBuilderError {
     #[snafu(display("List of parameter contains duplictates."))]
     DuplicateParameterNames,
 
@@ -14,15 +14,4 @@ pub enum ModelfunctionError {
     #[snafu(display("Subset contains parameters that are not in the full set"))]
     FunctionParameterNotInModel,
 
-}
-
-/// An error structure containing error variants related to
-/// the model and the set of model
-#[derive(Debug, Clone, Snafu, PartialEq)]
-#[snafu(visibility = "pub")]
-pub enum ModelBuilderError {
-    #[snafu(context(false))]
-    ModelfunctionError {
-        source : ModelfunctionError,
-    }
 }
